Split the async-client test file into a package
Break the monolithic tests/utils/test_github_async.py into tests/utils/github_async/:
- conftest.py: shared fixtures, payload factories, recording_transport, governed_client,
  make_zip, patch_signed_download, and the ENDPOINT_CALLS registry.
- test_client_core.py: construction, auth headers, context manager, timeout forwarding,
  PaginationData, and pagination mechanics.
- test_endpoints.py: per-endpoint success tests plus the registry-driven cross-cutting pair.
- test_models.py: model parsing contracts (already relocated earlier).
- test_download_artifact.py: redirect, token non-leak, zip-slip, signed-URL failures.
- test_rate_limiting.py: snapshot parsing, governor wiring, default construction, the retry
  suite, and the default-limiter logging tests folded in from test_github_async_defaults.py.

Coverage over ddev/utils/github_async is unchanged (430/430 executed lines).
